Create Tea Cozy 3
this is my third time trying to do Codecademy's Tea Cozy Project mostly on my own through trial and error and i just figured out myself how to move the nav bar up and down with css! I am soooo excited that i learned this through my own trial and error! I was so frustrated for so long... i welcome hints on how to continue on but i don't want this done for me.... but cheerleading and support would be appreciated i am so new to github because i have been so confused about it for so long i hope to learn more soon!